working spring project with data bound the 'spring way' @Kogurr
@Kogurr i used your Menu, MenuItem, and submenu items in order to bind a
Menu object to our model.
- I modified them slightly to fit with springs requirements but i think
that you will be able override your classes with my modified ones and
there will be no errors.
- The only thing i changed is added some getters and setters for a few
attributes and changed new ArrayList<>(3) to new ArrayList() on submenu
and menuItems declaration because it wouldn't work in spring how you had
it.

-I built a sample page "showMessage.jsp" located in WEB-INF/views that
shows how all the menu items can be accessed.

- now all you have to do is direct the Menucontrollers bindMenu method
to go to a page or another controller to create the pdf.
- To go to another page or request parameter(for a controller) you must
change
showMessage to the other page or request parameter in:

ModelAndView model = new ModelAndView("showMessage");

menu can be accessed by model.menu in a controller that takes (Model
model) as a request parameter

-  i can help with this process just let me know by email.

There is no implementation for editing an already created menu as of
right now but we should just try to get this working how it is before i
add that functionality.

I would use this maven shell that is already created because there is a
lot of xml configurations that  will most likely become broken if files
are just copied to a new shell.